+++ NEWS.md
@@ -2,6 +2,8 @@
 
 ### 2.4 (unreleased)
 
+- set the default message digest to sha256
+
 ### 2.3 (2022.03.06)
 
 **CRITICAL SECURITY VULNERABILITIES**